refactor: don't use private module names
