+++ setup.cfg
@@ -56,7 +56,7 @@ testing =
 
 [options.entry_points]
 console_scripts =
-    eric = molgenis.bbmri_eric._cli:run
+    eric = molgenis.bbmri_eric.cli:run
 
 [tool:pytest]
 # Specify command line options as you would do when invoking pytest directly.
