+++ src/molgenis/bbmri_eric/printer.py
@@ -1,5 +1,5 @@
-from molgenis.bbmri_eric._model import Node
 from molgenis.bbmri_eric.errors import EricError, EricWarning, ErrorReport
+from molgenis.bbmri_eric.model import Node
 
 
 class Printer:
