+++ src/molgenis/bbmri_eric/publisher.py
@@ -1,10 +1,10 @@
 from typing import List, Set
 
-from molgenis.bbmri_eric._enricher import Enricher
-from molgenis.bbmri_eric._model import Node, NodeData, QualityInfo, Table
-from molgenis.bbmri_eric._printer import Printer
 from molgenis.bbmri_eric.bbmri_client import EricSession
+from molgenis.bbmri_eric.enricher import Enricher
 from molgenis.bbmri_eric.errors import EricError, EricWarning
+from molgenis.bbmri_eric.model import Node, NodeData, QualityInfo, Table
+from molgenis.bbmri_eric.printer import Printer
 from molgenis.client import MolgenisRequestError
